requirements.txt
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0